+++ cypress/integration/signup.spec.js
@@ -19,7 +19,7 @@ describe("signup page", () => {
 
   it("Adds subpath for french pages", () => {
     cy.get('[data-cy="toggle-language-link"]').click();
-    cy.url().should("eq", "http://localhost:3000/fr/signup");
+    cy.url().should("eq", "http://localhost:3000/fr/inscription");
   });
 
   it("Menu appears on the page", () => {
@@ -105,6 +105,14 @@ describe("signup page", () => {
     cy.get('[id="languageEn"]').click();
     cy.get('[id="agreeToConditions"]').click();
     cy.get('[data-cy="signup-submit"]').click();
-    cy.url().should("contains", "/thankyou?e=s***%40****l.***&ref=signup");
+    cy.get('[data-cy="toggle-language-link"]').then(($el) => {
+      const text = $el.text();
+      if (text === "English") {
+        cy.url().should("contains", "/merci?e=s***%40****l.***&ref=signup");
+      }
+      if (text === "Français") {
+        cy.url().should("contains", "/thankyou?e=s***%40****l.***&ref=signup");
+      }
+    });
   });
 });