+++ CHANGELOG.md
@@ -7,6 +7,11 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+## Changed
+
+- Updated content on the experiments, about and sign up pages
+- Updated French path in French language
+
 ## [v1.1.2] - 2021-09-27
 
 ## Fixed